Remove "unitless" units
Real variables have units of '1'; logicals, strings, and integers that are
tracking non-physical quantities (like index of autotroph in grazing type) have
units of ''.

The init test now writes out units : '{units}', so empty strings show up as ''
and units of 1 are treated as a string instead of integer in
settings_metadata_check.py. (Also that python file did not set error_found if
the only differences between the JSON and Fortran metadata was in the units;
that bug has been fixed).