Only set Lig_prod(k)=0 when k>kmt
When I first created a branch with the iron forcing updates, the
call_compute_subroutines stand-alone test did not run; one of the issues was
that Lig_prod(k) was not defined for k > kmt. I thought I added a line of code
to set Lig_prod(k)=0 in that case, but inadvertantly was setting Lig_prod(k)=0
everywhere (overwriting the value computed earlier in the routine when k<=kmt).
The fix is to move the Lig_prod(k)=0 line into the else block of the "if (k <=
column_kmt)" statement