Updates to fix nvhpc compiler issues
To run the standalone tests on derecho, we want to use nvhpc/25.9 (this
includes a bug fix we reported where nvhpc/25.7 was not handling associate
statements correctly).

The compiler also rightly complained about an instance of a missing kind number
when using a real variable (although 1.0 is exactly representable so there
shouldn't be any difference between 1.0 and 1.0_r8).

Thirdly, I replaced a couple of where statements with the do-loop if-statement
combination to avoid dividing by zero. It is unfortunate that the compiler
doesn't handle this very well, but it is not a very invasive change.